+++ src/main/java/lotto/domain/Lotto.java
@@ -1,0 +1,42 @@
+package lotto.domain;
+
+import java.util.List;
+
+public class Lotto {
+    private static final int LOTTO_SIZE = 6;
+    private static final int MIN_NUMBER = 1;
+    private static final int MAX_NUMBER = 45;
+
+    private final List<Integer> numbers;
+
+    public Lotto(List<Integer> numbers) {
+        validate(numbers);
+        this.numbers = numbers;
+    }
+
+    private void validate(List<Integer> numbers) {
+        validateSize(numbers);
+        validateRange(numbers);
+        validateDuplicate(numbers);
+    }
+
+    private void validateSize(List<Integer> numbers){
+        if (numbers.size() != LOTTO_SIZE) {
+            throw new IllegalArgumentException("로또 번호는 6개여야 합니다.");
+        }
+    }
+
+    private void validateRange(List<Integer> numbers){
+        if(numbers.stream().anyMatch(num -> num < MIN_NUMBER || num > MAX_NUMBER)){
+            throw new IllegalArgumentException("로또 번호는 1~45 사이의 숫자여야 합니다.");
+        }
+    }
+
+    private void validateDuplicate(List<Integer> numbers){
+        long distinctCount = numbers.stream().distinct().count();
+        if (distinctCount != numbers.size()) {
+            throw new IllegalArgumentException("로또 번호는 중복될 수 없습니다.");
+        }
+    }
+
+}
